backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
psycopg2==2.9.10
pydotplus==2.0.2
pyparsing==3.2.0
sqlparse==0.5.2
typing_extensions==4.12.2
